requirements.txt
kivy[full]
kivymd
plyer
torch
torchvision
pillow
scikit-learn
numpy
matplotlib
pyrebase4